paystub컴포넌트 폴더 생성 및 계좌생성컴포넌트
