+++ client/src/app/components/PaystubPage/PaystubPreview.tsx
