+++ client/src/app/manager/paystub/page.tsx
@@ -1,14 +1,13 @@
 //import { useState } from 'react';
 "use client";
 import { useState } from "react";
-import GreenTop from "../../components/GreenTop";
+import GreenTop from "../../components/PaystubPage/GreenTop";
 import StaffSelect from "../../components/StaffSelect";
 import StaffSelectModal from "../../components/StaffSelectModal";
-import PaystubName from "../../components/PaystubName";
-import PaystubInput from "../../components/PaystubInput";
-import PaystubPreview from "../../components/PaystubPreview";
-import WorkingStatus from "../../components/WorkingStatus";
-import Navi from "../../components/managerNavi"
+import PaystubPreview from "../../components/PaystubPage/PaystubPreview";
+import WorkingStatus from "../../components/PaystubPage/WorkingStatus";
+import Navi from "../../components/managerNavi";
+import AccountAdd from "@/app/components/PaystubPage/AccountAdd";
 
 const Paystub = () => {
   const [showModal, setShowModal] = useState(false);
@@ -17,63 +16,37 @@ const Paystub = () => {
   //  const [isChecked, setIsChecked] = useState<boolean>(false);
   return (
     <>
-    <Navi />
-    <div className="">
-      
-      <GreenTop>직원선택</GreenTop>
-      <div className="bg-white p-3 m-5 flex justify-between">
-        <StaffSelect>홍길동</StaffSelect>
-        <button
-          className="text-sm font-bold hover:bg-gray-300 px-2"
-          onClick={() => setShowModal(true)}
-        >
-          지난달 급여내역 불러오기
-        </button>
+      <Navi />
+      <div className="">
+        <GreenTop>직원선택</GreenTop>
+        <div className="bg-white p-3 m-5 flex justify-between">
+          <StaffSelect>홍길동</StaffSelect>
+          <button
+            className="text-sm font-bold hover:bg-gray-300 px-2"
+            onClick={() => setShowModal(true)}
+          >
+            직원선택하기
+          </button>
+        </div>
+        {showModal && (
+          <StaffSelectModal onClose={() => setShowModal(false)}>
+            <StaffSelect>ddd</StaffSelect>
+            <StaffSelect>ddd</StaffSelect>
+            <StaffSelect>ddd</StaffSelect>
+            <StaffSelect>ddd</StaffSelect>
+            <StaffSelect>ddd</StaffSelect>
+            <StaffSelect>ddd</StaffSelect>
+            <StaffSelect>ddd</StaffSelect>
+            <StaffSelect>ddd</StaffSelect>
+          </StaffSelectModal>
+        )}
+        <GreenTop>계좌번호</GreenTop>
+        <AccountAdd />
+        <GreenTop>근태</GreenTop>
+        <WorkingStatus></WorkingStatus>
+        <GreenTop>지급내역</GreenTop>
+        <PaystubPreview></PaystubPreview>
       </div>
-      {showModal && (
-        <StaffSelectModal onClose={() => setShowModal(false)}>
-          <StaffSelect>ddd</StaffSelect>
-          <StaffSelect>ddd</StaffSelect>
-          <StaffSelect>ddd</StaffSelect>
-          <StaffSelect>ddd</StaffSelect>
-          <StaffSelect>ddd</StaffSelect>
-          <StaffSelect>ddd</StaffSelect>
-          <StaffSelect>ddd</StaffSelect>
-          <StaffSelect>ddd</StaffSelect>
-        </StaffSelectModal>
-      )}
-      <GreenTop>지급내역</GreenTop>
-      <PaystubPreview></PaystubPreview>
-      {/* <div className="bg-white p-3 m-5 flex-col ">
-        <div className="flex">
-          <PaystubName>기본금</PaystubName>
-          <PaystubInput></PaystubInput>
-        </div>
-        <div className="flex">
-          <PaystubName>통상시급</PaystubName>
-          <PaystubInput></PaystubInput>
-        </div>
-        <div className="flex">
-          <PaystubName>연장근로시간</PaystubName>
-          <PaystubInput></PaystubInput>
-        </div>
-        <div className="flex">
-          <PaystubName>
-            휴일근무시간<p className="text-sm">(8시간 초과)</p>
-          </PaystubName>
-          <PaystubInput></PaystubInput>
-        </div>
-        <div className="flex">
-          <PaystubName>
-            기타수당
-            <button className="bg-white w-7 h-7 ml-5">+</button>
-          </PaystubName>
-          <PaystubInput></PaystubInput>
-        </div>
-      </div> */}
-      <GreenTop>공제내역</GreenTop>
-      <WorkingStatus></WorkingStatus>
-    </div>
     </>
   );
 };
